dotfile generator: fish: use `fish_user_paths` instead of PATH
Assign user paths to the `fish_user_paths` variable instead of directly
prepending them to the PATH environment variable (fish(1) automatically
does this job instead).  This has the benefit of fixing a few bugs
related to multiple prepending of user paths to PATH on successive
fish(1) invocations (if they are prepended to the global variable of
that name) or prepending them to the wrong variable (if they are
prepended to the universal variable PATH instead of the environment
variable of that name, which is a global variable).

The `fish_user_paths` variable is created without explicit scope, so it
is assigned in the global scope when the script is sourced during
fish(1) startup.