gen: convert all definitions not to modify the fish config directory
Use universal variables for abbreviations and environment variable
definitions so that the fish_variables file won't be modified.  Don't
use funcsave for functions so that definitions in the functions
directory won't be modified.

Move the scripts appropriately to the fish config directory.  Rework the
generators for git abbreviations/aliases to use global variables and not
to use funcsave.