lvim: add some essential plugins; add Git diff and LSP diagnostic stats to the statusline
Add `symbols-outline.nvim` which displays an outline of symbols in the
current document. Add `lsp_signature.nvim` which displays the signature
of a function while you're typing its invocation.  Add
`todo-comments.nvim` which highlights TODO comments. Add `nvim-navic`
which displays the name of the current code context in the statusline.